+++ src/utils/nodejs.rs
@@ -23,6 +23,13 @@ pub fn arch() -> &'static str {
     }
 }
 
+pub fn bin_dir() -> PathBuf {
+    #[cfg(windows)]
+    return install_path();
+    #[cfg(not(windows))]
+    return install_path().join("bin");
+}
+
 pub fn current_version() -> String {
     match utils::process::command_output("node", &["--version"]) {
         Ok(output) => {
@@ -38,9 +45,9 @@ pub fn current_version() -> String {
 
 pub fn has_node() -> bool {
     #[cfg(windows)]
-    let exe_path = install_path().join("bin").join("node.exe");
+    let exe_path = bin_dir().join("node.exe");
     #[cfg(not(windows))]
-    let exe_path = install_path().join("bin").join("node");
+    let exe_path = bin_dir().join("node");
 
     exe_path.is_file()
 }
@@ -78,7 +85,7 @@ pub fn has_yarn() -> bool {
     }
 }
 
-fn install_path() -> PathBuf {
+pub fn install_path() -> PathBuf {
     utils::env::home_dir().join(".local").join("node")
 }
 
@@ -104,6 +111,13 @@ pub fn latest_version() -> String {
         .unwrap();
 
     String::from(latest_release.version.as_str().trim())
+}
+
+pub fn lib_dir() -> PathBuf {
+    #[cfg(windows)]
+    return install_path();
+    #[cfg(not(windows))]
+    return install_path().join("lib");
 }
 
 pub fn os() -> &'static str {